components/toxic-predictor/requirements.txt
scipy
scikit-learn
joblib
google-cloud-storage
pandas
numpy
requests
flask